requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.1
pkg-resources==0.0.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
